Move get_source_mac_addr
