+++ boreas/util.c
@@ -19,6 +19,12 @@
 
 #include "util.h"
 
+#include <errno.h>
+#include <glib.h>
+#include <ifaddrs.h>          /* for getifaddrs() */
+#include <net/if.h>           /* for if_nametoindex() */
+#include <netpacket/packet.h> /* for sockaddr_ll */
+#include <stdlib.h>
 /**
  * @brief Checksum calculation.
  *
@@ -56,4 +62,57 @@ in_cksum (uint16_t *addr, int len)
   sum += (sum >> 16);                 /* add carry */
   answer = ~sum;                      /* truncate to 16 bits */
   return (answer);
+}
+
+/**
+ * @brief Get the source mac address of the given interface
+ * or of the first non lo interface.
+ *
+ * @param interface Interface to get mac address from or NULL if first non lo
+ * interface should be used.
+ * @param[out]  mac Location where to store mac address.
+ *
+ * @return 0 on success, -1 on error.
+ */
+int
+get_source_mac_addr (char *interface, uint8_t *mac)
+{
+  struct ifaddrs *ifaddr = NULL;
+  struct ifaddrs *ifa = NULL;
+  int interface_provided = 0;
+
+  if (interface)
+    interface_provided = 1;
+
+  if (getifaddrs (&ifaddr) == -1)
+    {
+      g_debug ("%s: getifaddr failed: %s", __func__, strerror (errno));
+      return -1;
+    }
+  else
+    {
+      for (ifa = ifaddr; ifa != NULL; ifa = ifa->ifa_next)
+        {
+          if ((ifa->ifa_addr) && (ifa->ifa_addr->sa_family == AF_PACKET)
+              && !(ifa->ifa_flags & (IFF_LOOPBACK)))
+            {
+              if (interface_provided)
+                {
+                  if (g_strcmp0 (interface, ifa->ifa_name) == 0)
+                    {
+                      struct sockaddr_ll *s =
+                        (struct sockaddr_ll *) ifa->ifa_addr;
+                      memcpy (mac, s->sll_addr, 6 * sizeof (uint8_t));
+                    }
+                }
+              else
+                {
+                  struct sockaddr_ll *s = (struct sockaddr_ll *) ifa->ifa_addr;
+                  memcpy (mac, s->sll_addr, 6 * sizeof (uint8_t));
+                }
+            }
+        }
+      freeifaddrs (ifaddr);
+    }
+  return 0;
 }
